Update github New Issue workflow to mimic Remix
