+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -11,18 +11,21 @@ body:
 
         Do you need some help?
         ======================
-        The issue tracker is meant for feature requests and bug reports only. This isn't the best place for
-        support or usage questions. Questions here don't have as much visibility as they do elsewhere. Before
-        you ask a question, here are some resources to get help first:
+        The issue tracker is meant for bug reports only. This isn't the best
+        place for support or usage questions. Questions here don't have as much
+        visibility as they do elsewhere. Before you ask a question, here are
+        some resources to get help first:
 
         - Read the docs: https://reactrouter.com
         - Check out the list of frequently asked questions: https://reactrouter.com/start/faq
         - Explore examples: https://reactrouter.com/start/examples
-        - Look for/ask questions on Stack Overflow: https://stackoverflow.com/questions/tagged/react-router
         - Ask in chat: https://discord.gg/6RyV8n8yyM
+        - Look for/ask questions on Stack Overflow: https://stackoverflow.com/questions/tagged/react-router
 
-        ### Test Case Starter:
-        https://stackblitz.com/github/remix-run/react-router/tree/main/examples/basic?file=src/App.tsx
+        ### Test Case Starters:
+
+        * (Using `<RouterProvider>`)[https://stackblitz.com/github/remix-run/react-router/tree/main/examples/data-router?file=src/App.tsx]
+        * (Using `<BrowserRouter>`)[https://stackblitz.com/github/remix-run/react-router/tree/main/examples/basic?file=src/App.tsx]
   - type: input
     attributes:
       label: What version of React Router are you using?
