--- .github/ISSUE_TEMPLATE/feature_request.yml
@@ -1,32 +1,0 @@
-name: 💡 Feature Request
-description: I want to add something to React Router.
-title: "[Feature]: "
-labels:
-  - feature
-body:
-  - type: markdown
-    attributes:
-      value: |
-        Thank you for contributing!
-
-        Do you need some help?
-        ======================
-        The issue tracker is meant for feature requests and bug reports only. This isn't the best place for 
-        support or usage questions. Questions here don't have as much visibility as they do elsewhere. Before 
-        you ask a question, here are some resources to get help first:
-
-        - Read the docs: https://reactrouter.com
-        - Check out the list of frequently asked questions: https://reactrouter.com/start/faq
-        - Explore examples: https://reactrouter.com/start/examples
-        - Look for/ask questions on Stack Overflow: https://stackoverflow.com/questions/tagged/react-router
-        - Ask in chat: https://discord.gg/6RyV8n8yyM
-  - type: textarea
-    attributes:
-      label: What is the new or updated feature that you are suggesting?
-    validations:
-      required: true
-  - type: textarea
-    attributes:
-      label: Why should this feature be included?
-    validations:
-      required: true
